+++ vms/client/nx_vms_client_desktop/src/nx/vms/client/desktop/event_rules/models/detectable_object_type_model.cpp
@@ -38,7 +38,7 @@ public:
 
     void updateData(const std::vector<ObjectType*>& types)
     {
-        for (int i = 0; i < types.size(); ++i)
+        for (int i = 0; i < (int) types.size(); ++i)
         {
             ObjectType* type = types[i];
             data.insert(type, i);
